+++ src/webserver/database/sqlpage_functions/functions.rs
@@ -22,7 +22,7 @@ super::function_definition_macro::sqlpage_functions! {
     environment_variable(name: Cow<str>);
     exec((&RequestInfo), program_name: Cow<str>, args: Vec<Cow<str>>);
 
-    fetch(http_request: SqlPageFunctionParam<super::http_fetch_request::HttpFetchRequest<'_>>);
+    fetch((&RequestInfo), http_request: SqlPageFunctionParam<super::http_fetch_request::HttpFetchRequest<'_>>);
 
     hash_password(password: Option<String>);
     header((&RequestInfo), name: Cow<str>);
@@ -129,12 +129,12 @@ async fn exec<'a>(
 }
 
 async fn fetch(
+    request: &RequestInfo,
     http_request: super::http_fetch_request::HttpFetchRequest<'_>,
 ) -> anyhow::Result<String> {
     use awc::http::Method;
-    let client = awc::Client::builder()
-        .add_default_header((awc::http::header::USER_AGENT, env!("CARGO_PKG_NAME")))
-        .finish();
+    let client = make_http_client(&request.app_state.config);
+
     let method = if let Some(method) = http_request.method {
         Method::from_str(&method)?
     } else {
@@ -172,6 +172,27 @@ async fn fetch(
     let response_str = String::from_utf8(body)?;
     log::debug!("Fetch response: {response_str}");
     Ok(response_str)
+}
+
+fn make_http_client(config: &crate::app_config::AppConfig) -> awc::Client {
+    let connector = if config.system_root_ca_certificates {
+        let mut roots = rustls::RootCertStore::empty();
+        for cert in rustls_native_certs::load_native_certs().expect("could not load platform certs")
+        {
+            roots.add(cert).unwrap();
+        }
+        let tls_conf = rustls::ClientConfig::builder()
+            .with_root_certificates(roots)
+            .with_no_client_auth();
+
+        awc::Connector::new().rustls_0_22(std::sync::Arc::new(tls_conf))
+    } else {
+        awc::Connector::new()
+    };
+    awc::Client::builder()
+        .connector(connector)
+        .add_default_header((awc::http::header::USER_AGENT, env!("CARGO_PKG_NAME")))
+        .finish()
 }
 
 pub(crate) async fn hash_password(password: Option<String>) -> anyhow::Result<Option<String>> {